feat: add auth to api routes

+++ src/app/api/transaksi/route.ts
@@ -21,6 +21,7 @@ export const GET = async (req: NextRequest) => {
 
 export const POST = async (req: Request) => {
     const session = await getServerSession();
+    if (!session?.user?.role?.includes("KASIR")) return NextResponse.json({ status: 403, message: "forbidden" }, { status: 403 });
 
     try {
         const data = await req.formData();
@@ -37,18 +38,25 @@ export const POST = async (req: Request) => {
     }
 }
 
-export const PATCH = async (req: NextRequest) => {
+export const PUT = async (req: NextRequest) => {
     const queryParams = req.nextUrl.searchParams;
-    const session = await getServerSession();
     const id = Number(queryParams?.get('id'));
+
+    const session = await getServerSession();
+    if (!session?.user?.role?.includes("KASIR")) return NextResponse.json({ status: 403, message: "forbidden" }, { status: 403 });
 
     if (!id) {
         return NextResponse.json({ status: 403, message: "forbidden" }, { status: 403 });
     }
 
     try {
+        const findOne = await prisma.transaksi.findUnique({ where: { id } });
+        if (!findOne) {
+            return NextResponse.json({ status: 404, message: "not found" }, { status: 404 });
+        }
+        
         const data = await req.formData();
-        const idUser = Number(session?.user?.id); // Get from session
+        const idUser = Number(session?.user?.id); // Get idKasir from session
         const nomorPesanan = data.get('nomorPesanan') as string;
         const totalPesanan = Number(data.get('totalPesanan') as string);
         const idStand = Number(data.get('idStand') as string);
@@ -64,13 +72,22 @@ export const DELETE = async (req: NextRequest) => {
     const queryParams = req.nextUrl.searchParams;
     const id = Number(queryParams?.get("id"));
 
+    const session = await getServerSession();
+    if (!session?.user?.role?.includes("KASIR")) return NextResponse.json({ status: 403, message: "forbidden" }, { status: 403 });
+
     if (!id) {
         return NextResponse.json({ status: 403, message: "forbidden" }, { status: 403 });
     }
 
     try {
+        const findOne = await prisma.transaksi.findUnique({ where: { id } });
+        if (!findOne) {
+            return NextResponse.json({ status: 404, message: "not found" }, { status: 404 });
+        }
+
         await prisma.transaksi.delete({ where: { id } });
         return NextResponse.json({ status: 200, message: 'success' }, { status: 200 });
+
     } catch {
         return NextResponse.json({ status: 500, message: 'internal server error' });
     }
